createSliceMachineManagerMSWHandler in __setup__
