+++ packages/slice-machine/test/components/CreateCustomTypeModal.test.tsx
@@ -6,10 +6,6 @@ import { render, fireEvent, act } from "../__testutils__";
 import CreateCustomTypeModal from "../../components/Forms/CreateCustomTypeModal";
 import { ModalKeysEnum } from "../../src/modules/modal/types";
 import { rest } from "msw";
-import { createTestPlugin } from "test/__testutils__/createTestPlugin";
-import { createTestProject } from "test/__testutils__/createTestProject";
-import { createSliceMachineManager } from "@slicemachine/manager";
-import { createSliceMachineManagerMSWHandler } from "@slicemachine/manager/test";
 
 vi.mock("next/router", () => require("next-router-mock"));
 global.console = { ...global.console, error: vi.fn() };
@@ -24,23 +20,6 @@ describe("CreateCustomTypeModal", () => {
   document.body.appendChild(div);
 
   test("when a slice is created the tracker should be called", async (ctx) => {
-    const adapter = createTestPlugin();
-    const cwd = await createTestProject({ adapter });
-    const manager = createSliceMachineManager({
-      nativePlugins: { [adapter.meta.name]: adapter },
-      cwd,
-    });
-
-    await manager.telemetry.initTelemetry();
-    await manager.plugins.initPlugins();
-
-    ctx.msw.use(
-      createSliceMachineManagerMSWHandler({
-        url: "http://localhost:3000/_manager",
-        sliceMachineManager: manager,
-      })
-    );
-
     const trackingSpy = vi.fn<Parameters<Parameters<typeof rest.post>[1]>>(
       (_req, res, ctx) => res(ctx.json({}))
     );
